requirements.txt
# Core dependencies
# PyQt6==6.6.1
mysql-connector-python==8.2.0
twilio==8.11.0
openpyxl==3.1.2
# pandas==2.1.4
argon2-cffi==23.1.0
python-dotenv==1.0.0
schedule==1.2.0
Pillow==10.2.0

# Development dependencies
pylint==3.0.3
black==23.12.1
pytest==7.4.4
